utils/gemini_utils.py
# 이 파일은 가상의 Gemini 2.0 Flash API를 가정하여 작성되었습니다.
# 실제 Gemini API가 출시되면 그에 맞게 코드를 수정해야 합니다.
import os
import io
import soundfile as sf
import sounddevice as sd

# Gemini API Key (환경 변수에서 가져오기)
GEMINI_API_KEY = os.environ.get("GEMINI_API_KEY")

# (가정) Gemini API 클라이언트 
class GeminiClient:
    def __init__(self, api_key):
        self.api_key = api_key

    def generate_response(self, audio_content, instructions, conversation_history):
        """
        Gemini API에 오디오 파일을 보내고 응답을 생성합니다. (가정)

        Args:
            audio_content (bytes): 오디오 파일의 바이너리 데이터입니다.
            instructions (dict): 사용자가 설정한 지시사항입니다.
            conversation_history (list): 이전 대화 기록입니다.

        Returns:
            str: Gemini API의 텍스트 응답입니다.
        """
        # (가정) Gemini API에 오디오 파일과 지시사항, 대화 기록을 보내고 응답을 받습니다.
        # 이 부분은 실제 Gemini API의 스펙에 맞게 구현되어야 합니다.
        # 현재는 더미 데이터로 응답을 시뮬레이션합니다.
        print(f"Received audio content: {len(audio_content)} bytes")
        print(f"Instructions: {instructions}")
        print(f"Conversation history: {conversation_history}")

        # 더미 응답 생성
        dummy_response = f"This is a dummy response from Gemini 2.0 Flash. (Instructions: {instructions})"

        return dummy_response

# (가정) Gemini API 클라이언트 생성
gemini_client = GeminiClient(GEMINI_API_KEY)

def generate_gemini_response(audio_file_path, instructions, conversation_history=[]):
    """
    Gemini API를 사용하여 오디오 파일에 대한 응답 생성 (음성 입력 처리 가정)

    Args:
        audio_file_path (str): 오디오 파일 경로
        instructions (dict): 사용자 지시사항
        conversation_history (list): 대화 기록

    Returns:
        str: Gemini의 텍스트 응답
    """
    try:
        with io.open(audio_file_path, "rb") as audio_file:
            audio_content = audio_file.read()

        # Gemini API 호출 (가정: 음성 입력 지원)
        response_text = gemini_client.generate_response(audio_content, instructions, conversation_history)

        return response_text

    except Exception as e:
        print(f"Error in generate_gemini_response: {e}")
        return "Sorry, I encountered an error while processing your request."

def synthesize_and_play_text(text, repeat_count=1, output_filename="gemini_response.wav"):
    """
    텍스트를 음성으로 변환 후 재생 (여기서는 Google Cloud Text-to-Speech 사용)

    Args:
        text (str): 변환할 텍스트
        repeat_count (int): 반복 재생 횟수
        output_filename (str): 출력 파일명
    """
    try:
        from google.cloud import texttospeech_v1 as texttospeech

        # Google Cloud Text-to-Speech 클라이언트 초기화
        tts_client = texttospeech.TextToSpeechClient()

        synthesis_input = texttospeech.SynthesisInput(text=text)
        voice = texttospeech.VoiceSelectionParams(
            language_code="en-US", ssml_gender=texttospeech.SsmlVoiceGender.NEUTRAL
        )
        audio_config = texttospeech.AudioConfig(
            audio_encoding=texttospeech.AudioEncoding.LINEAR16, # WAV
            speaking_rate=1.1
        )

        response = tts_client.synthesize_speech(
            input=synthesis_input, voice=voice, audio_config=audio_config
        )

        # 음성 데이터를 파일로 저장
        with open(output_filename, "wb") as out:
            out.write(response.audio_content)
            print(f'Audio content written to file "{output_filename}"')

        # 반복 재생
        for _ in range(repeat_count):
            data, samplerate = sf.read(output_filename)
            sd.play(data, samplerate)
            sd.wait()

    except Exception as e:
        print(f"Error in synthesize_and_play_text: {e}")